+++ source/models/AudioSendersModel.hpp
@@ -109,14 +109,29 @@ private:
 
         void ravenna_sender_configuration_updated (rav::Id sender_id, const rav::RavennaSender::Configuration& configuration) override;
 
-        void prepareInput (rav::AudioFormat inputFormat);
-        void processBlock (const rav::AudioBufferView<const float>& inputBuffer, uint32_t timestamp) const;
+        void prepareInput (rav::AudioFormat inputFormat, uint32_t maxNumFramesPerBlock);
+        void processBlock (const rav::AudioBufferView<const float>& inputBuffer, uint32_t rtpTimestamp, rav::ptp::Timestamp ptpTimestamp);
+        void resetInput();
 
     private:
+        struct RealtimeSharedContext
+        {
+            uint32_t inputSampleRate {};
+            uint32_t targetSampleRate {};
+            std::unique_ptr<Resample, decltype (&resampleFree)> resampler { nullptr, &resampleFree };
+            rav::AudioBuffer<float> resampleBuffer;
+            std::optional<uint32_t> rtpTimestamp {}; // Used only when resampler is active
+        };
         [[maybe_unused]] AudioSendersModel& owner_;
         rav::Id senderId_;
         SenderState state_;
+        uint32_t maxNumFramesPerBlock_ = 0;
+
+        rav::RealtimeSharedObject<RealtimeSharedContext> realtimeSharedContext_;
+
         MessageThreadExecutor executor_; // Keep last so that it's destroyed first
+
+        void updateRealtimeSharedContext();
     };
 
     struct RealtimeSharedContext
@@ -132,6 +147,7 @@ private:
     // Accessed by audio thread:
     std::optional<uint32_t> rtpTs_ {};
     rav::AudioFormat deviceFormat_;
+    uint32_t maxNumFramesPerBlock_ = 0;
     rav::RealtimeSharedObject<RealtimeSharedContext> realtimeSharedContext_;
     rav::AudioBuffer<float> resamplerBuffer_ {};
     std::unique_ptr<Resample, decltype (&resampleFree)> resampler_ { nullptr, &resampleFree };